% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speech_data.R
\name{get_speech_data}
\alias{get_speech_data}
\title{Return all speeches of a speaker}
\usage{
get_speech_data(speaker, from = NULL, until = NULL, show_progress = TRUE)
}
\arguments{
\item{speaker}{Name of speaker - (Hiragana is acceptable for the member's name). Partial match search. If multiple single-byte spaces (U+0020) are specified as delimiters, the specified words will be OR-searched.}

\item{from}{Optional start date of search of the form YYYY-MM-DD}

\item{until}{Optional end date of search of the form YYYY-MM-DD}

\item{show_progress}{Boolean to show a progress bar; TRUE by default}
}
\description{
Return all speeches of a speaker
}
\examples{
# Download all speech data for Shinzo Abe for the year 2006
abe_speeches_2006 <- get_speech_data(
  speaker = "安倍 晋三",
  from = "2006-01-01",
  until = "2006-12-31"
)

}
