% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kokkai_url.R
\name{kokkai_call}
\alias{kokkai_call}
\title{Get Kokkai API request as R list}
\usage{
kokkai_call(unit = "speech", ...)
}
\arguments{
\item{unit}{Unit of request, must be one of "speech", "meeting", or "meetingList". Default value is "speech".}

\item{...}{Other parameters to pass to kokkai_url()}
}
\value{
R list representation of JSON
}
\description{
Get Kokkai API request as R list
}
\seealso{
[kokkai_url()] for full list of arguments
}
